Moved CsItem to Modules/Entities/Constants
